Seperates camera concretes from types
