+++ src/pyrite/__init__.py
@@ -12,7 +12,8 @@ from .game import (  # noqa:F401
     get_render_manager,
     get_renderer,
 )
-from .types.camera import Camera, ChaseCamera  # noqa:F401
+
+from .camera import Camera, ChaseCamera  # noqa:F401
 from .types.entity import Entity  # noqa:F401
 from .types.renderable import Renderable  # noqa:F401
 from .types.enums import RenderLayers, AnchorPoint  # noqa: F401
