+++ src/pyrite/camera/__init__.py
@@ -1,0 +1,2 @@
+from .camera import Camera  # noqa:F401
+from .chase_camera import ChaseCamera  # noqa:F401
