+++ src/pyrite/types/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Protocol, runtime_checkable, TYPE_CHECKING
 
-from .camera import CameraBase, Camera, ChaseCamera  # noqa: F401
+from .camera import CameraBase  # noqa: F401
 from .entity import Entity  # noqa: F401
 from .enums import RenderLayers, AnchorPoint, Layer, Anchor  # noqa: F401
 from .renderable import Renderable  # noqa: F401
